Add driver quit() after occurred error, and fix FoodLoading Error
1.
driver quit() needed after error occurred
when the error occurred the driver didn't close so the chrome driver use
the memory. It cause the lack of memory /dev/shm

2. FoodLoadingError
when load overEasyEggPatty  the extracted temp String was "140
" with the a lots space.
so using trim(), delete the space and load worked in the test